Remove redundant /estimate page and fix Book Now flow to launch payment modal directly
